Renames header file since Github is not understanding lowercase
